+++ examples/with-emailverification-then-password-thirdpartyemailpassword/src/Home/index.js
@@ -10,21 +10,12 @@ export default function Home() {
     const sessionContext = useSessionContext();
     const navigate = useNavigate();
 
-    useEffect(() => {
-        if (
-            !sessionContext.loading &&
-            sessionContext.invalidClaims.find((err) => err.validatorId === RealPasswordClaim.id)
-        ) {
-            navigate("/set-password");
-        }
-    }, [sessionContext, navigate]);
-
     async function logoutClicked() {
         await signOut();
         navigate("/auth");
     }
 
-    if (sessionContext.loading || sessionContext.invalidClaims.length > 0) {
+    if (sessionContext.loading) {
         return null;
     }