+++ examples/with-thirdpartyemailpassword-2fa-passwordless/src/secondFactorClaim.tsx
@@ -5,4 +5,5 @@ export const SecondFactorClaim = new BooleanClaim({
     refresh: async () => {
         // This is something we have no way of refreshing, so this is a no-op
     },
+    onFailureRedirection: () => "/second-factor",
 });