+++ lib/ts/recipe/session/components/themes/translations.ts
@@ -1,0 +1,11 @@
+import { defaultTranslationsCommon } from "../../../../translation/translations";
+
+export const defaultTranslationsSession = {
+    en: {
+        ...defaultTranslationsCommon.en,
+
+        ACCESS_DENIED: "Access denied",
+        GO_BACK: "Go back",
+        LOGOUT: "Log out",
+    },
+};